solving comments from Manuel in PR #4
