+++ Features/UI/Careers.feature
@@ -6,7 +6,7 @@ Background: User navigate to EPAM and click on accept cookies
 
 Scenario Outline: User performs a career search and verifies job details
 	Given the user navigates to the Careers page
-	When the user searches for a career with "<searchText>", in "<locationValue>", and with <locationModality> modality
+	When the user searches for a career with "<searchText>", in "<locationValue>", and with "<locationModality>" modality
 	And the user applies and views the job from the last section
 	Then the job details should contain "<searchText>"
