+++ Pages/Article/ArticlePage.locator.cs
@@ -4,6 +4,6 @@ namespace DotnetTaskSeleniumNunit.Pages.Article;
 
 internal partial class ArticlePage
 {
-    private readonly By _artivcleTitle = By.CssSelector("main p:first-of-type[style],div.content-container>div.section:first-of-type p[style]");
-    private IWebElement ArticleTitle => _driver.FindElement(_artivcleTitle);
+    private readonly By _articleTitle = By.CssSelector("main p:first-of-type[style],div.content-container>div.section:first-of-type p[style]");
+    private IWebElement ArticleTitle => _driver.FindElement(_articleTitle);
 }
