+++ Pages/Navigation/NavigationBar.message.cs
@@ -4,6 +4,6 @@ internal partial class NavigationBar
 {
     private const string _errorClickingTopLink = "Error trying to search and click the link from top bar";
     private const string _infoCookiesSkipped = "No Cookies prompt found, skipping";
-    private const string _ErrorHoverLink = "Error while trying to hover on a link element";
-    private const string _ErrorClickSubItemFromHoveredLink = "Failed to click on a sub item from the hovered link";
+    private const string _errorHoverLink = "Error while trying to hover on a link element";
+    private const string _errorClickSubItemFromHoveredLink = "Failed to click on a sub item from the hovered link";
 }
