+++ Pages/Services/ServicesPage.locator.cs
@@ -1,13 +1,12 @@
 ﻿using NUnit.Framework;
 using System.ComponentModel.DataAnnotations;
 using OpenQA.Selenium;
-using OpenQA.Selenium.BiDi.Modules.BrowsingContext;
 
 namespace DotnetTaskSeleniumNunit.Pages.About;
 
 internal partial class ServicesPage
 {
-    private readonly By _LabelOurRelatedExpertise = By.XPath("//span[@class='museo-sans-light' and contains(text(),'Our Related Expertise')]");
+    private readonly By _labelOurRelatedExpertise = By.XPath("//span[@class='museo-sans-light' and contains(text(),'Our Related Expertise')]");
 
-    private IWebElement OurRelatedExpertiseLabel => _driver.FindElement(_LabelOurRelatedExpertise);
+    private IWebElement OurRelatedExpertiseLabel => _driver.FindElement(_labelOurRelatedExpertise);
 }
